separate bussiness logic

+++ services/categories/services/category.services.ts
@@ -2,9 +2,11 @@ import Category from '../models/category.model';
 import CategoryType from '../interfaces/Category';
 import { Op } from 'sequelize';
 import axios from 'axios';
+import AppError from '../errorHandler/AppError';
 
-export const createCategory = (name: string) => {
-    return Category.findOrCreate({
+export const createCategory = async (name: string, token: string | undefined) => {
+    await checkAuth(token);
+    const [category, created] = await Category.findOrCreate({
         where: {
             name: {
                 [Op.iLike]: name,
@@ -14,39 +16,70 @@ export const createCategory = (name: string) => {
             name,
         },
     });
+
+    if (!created) {
+        throw new AppError(`Category with name ${name} is already exist`, 400, 'bad-request');
+    }
+
+    return category;
 };
 
-export const findCategoryById = (id: string) => {
-    return Category.findOne({
+export const findCategoryById = async (categoryId: string) => {
+    const category = await Category.findOne({
         where: {
-            id,
+            id: categoryId,
         },
     });
+
+    if (!category) {
+        throw new AppError(`Category with id ${categoryId} not found`, 404, 'not-found');
+    }
+
+    return category;
 };
 
 export const findAllCategories = () => {
     return Category.findAll();
 };
 
-export const updateCategoryById = (updatedCategory: CategoryType, id: string) => {
-    return Category.update(updatedCategory, {
+export const updateCategoryById = async (
+    updatedCategory: CategoryType,
+    categoryId: string,
+    token: string | undefined,
+) => {
+    await checkAuth(token);
+
+    const category = await findCategoryById(categoryId);
+
+    if (!category) {
+        throw new AppError(`Category with id ${categoryId} not found`, 404, 'not-found');
+    }
+
+    const [_, _updatedCategory] = await Category.update(updatedCategory, {
         where: {
-            id,
+            id: categoryId,
         },
         returning: true,
     });
+
+    return _updatedCategory;
 };
 
-export const deleteCategoryById = (id: string) => {
-    return Category.destroy({
+export const deleteCategoryById = async (categoryId: string, token: string | undefined) => {
+    await checkAuth(token);
+    const category = await findCategoryById(categoryId);
+
+    if (!category) {
+        throw new AppError(`Category with id ${categoryId} not found`, 404, 'not-found');
+    }
+
+    Category.destroy({
         where: {
-            id,
+            id: categoryId,
         },
     });
-};
 
-export const getCategoryById = (id: string) => {
-    return id;
+    return category;
 };
 
 export const checkAuth = async (token: string | undefined) => {
@@ -58,12 +91,13 @@ export const checkAuth = async (token: string | undefined) => {
         };
     }
     const response = await axios.post(
-        `http://services_authentication_1:8081/api/v1/auth/check-auth`,
+        `http://apigateway:8080/api/v1/auth/check-auth`,
         {},
         {
             headers: headersConfig,
         },
     );
+    console.log(response);
     const uid = response.data.data.uid;
     return uid;
 };
